+++ graph_attention_student/examples/vgd_multitask.py
@@ -1,6 +1,22 @@
 """
 This example shows how to train a MEGAN model for multitask graph regression application based on an
-existing "visual graph dataset"
+existing "visual graph dataset".
+
+In this example the VGD "organic_solvents" is used. This datasets consists of a few thousand molecules,
+which are annotated with 4 target values, representing measurements of solubility within 4 different
+organic solvents: water, ethanol, benzene, acetone. For this dataset it is important to note that not
+every element is annotated with all values. Most of the time, the target value vectors contain "NaN" fields
+which means that the downstream processing pipeline and graph neural network are able to handle this.
+
+**CHANGELOG**
+
+0.1.0 - 16.12.2022 - Initial version
+
+0.2.0 - 30.12.2022 - Additional artifacts have been added: (1) After loading of the dataset a PDF with
+information about the target value distribution of the dataset is now created. This is supposed to help
+for deciding on appropriate REGRESSION_REFERENCE and REGRESSION_LIMITS values. (2) After the evaluation
+step, a PDF with the test set results is created containing a regression plot and the R2 value for each of
+the targets.
 """
 import os
 import warnings
@@ -40,16 +56,19 @@ from graph_attention_student.models import MultiAttentionStudent
 from graph_attention_student.models import Megan
 from graph_attention_student.visualization import plot_node_importances
 from graph_attention_student.visualization import plot_edge_importances
+from graph_attention_student.visualization import plot_regression_fit
 
 np.set_printoptions(precision=2)
 
-
+VERSION = '0.2.0'
 SHORT_DESCRIPTION = (
-    'This example shows how to train a MEGAN model for multitask graph regression application based on an existing '
-    'visual graph dataset.'
+    'This example shows how to train a MEGAN model for multitask graph regression application based on an '
+    'existing visual graph dataset.'
 )
 
 # == DATASET PARAMETERS ==
+# These parameters are used to specify the dataset to be used for the training as well as additional
+# properties of the dataset such as the train test split for example.
 
 # The name of the visual graph dataset to use for this experiment.
 VISUAL_GRAPH_DATASET_PATH = os.path.expanduser('~/.visual_graph_datasets/datasets/organic_solvents')
@@ -61,9 +80,10 @@ NUM_TARGETS = 4
 # Whether the dataset already contains importance (explanation) ground truth annotations.
 # most of the time this will most likely not be the case
 HAS_IMPORTANCES: bool = False
-# IF the dataset includes global graph attributes and if they are supposed to be used in the training process, this
-# flag has to be set to True.
+# IF the dataset includes global graph attributes and if they are supposed to be used in the training
+# process, this flag has to be set to True.
 HAS_GRAPH_ATTRIBUTES: bool = True
+HAS_GRAPH_ATTRIBUTES = False
 # The ratio of the test set to be used as examples for the visualization of the explanations
 EXAMPLES_RATIO: float = 0.2
 # The string names of the target values in the order in which they appear in the dataset as well
@@ -76,10 +96,11 @@ TARGET_NAMES = [
 ]
 
 # == MODEL PARAMETERS ==
+# These paremeters can be used to configure the model
 
 # This list defines how many graph convolutional layers to configure the network with. Every element adds
 # one layer. The numbers themselves are the hidden units to be used for those layers.
-UNITS = [20, 20, 20, 20]
+UNITS = [20, 20, 20, 20, 20]
 
 # The dropout rate which is applied after EVERY graph convolutional layer of the network. Especially for
 # large networks (>20 hidden units and multiple importance channels, this dropout proves very useful)
@@ -124,12 +145,14 @@ REGRESSION_LIMITS = [
 # depth of that MLP. Each element adds a layer and the values themselves dictate the hidden values.
 # In the regression case a final entry with "1" hidden layer as the output layer will implicitly be
 # added.
-FINAL_UNITS = [50, 20, NUM_TARGETS]
+FINAL_UNITS = [100, 50, NUM_TARGETS]
 
 # This dropout is applied after EVERY layer in the final MLP. Using this should not be necessary
-FINAL_DROPOUT = 0.0
+FINAL_DROPOUT = 0.1
 
 # == TRAINING PARAMETERS ==
+# These parameters are hyperparameters that control the training process itself.
+
 DEVICE = 'gpu:0'
 LEARNING_RATE = 0.001
 BATCH_SIZE = 64
@@ -139,6 +162,8 @@ EPOCHS = 200
 LOG_STEP = 5
 LOG_STEP_EVAL = 1000
 METRIC_KEY = 'mean_squared_error'
+COLOR_PRIMARY = 'gray'
+COLOR_SECONDARY = 'black'
 
 # == EXPERIMENT PARAMETERS ==
 DEBUG = True
@@ -155,6 +180,7 @@ with Skippable(), (e := Experiment(base_path=BASE_PATH, namespace=NAMESPACE, glo
         metadata_contains_index=True
     )
     dataset_indices = list(sorted(index_data_map.keys()))
+    dataset_length = len(index_data_map)
     dataset: t.List[tc.GraphDict] = []
     for index in dataset_indices:
         data = index_data_map[index]
@@ -170,6 +196,66 @@ with Skippable(), (e := Experiment(base_path=BASE_PATH, namespace=NAMESPACE, glo
     test_indices = [index for index in dataset_indices if index not in train_indices]
     e['train_indices'] = train_indices
     e['test_indices'] = test_indices
+
+    # ~ Visualize Dataset Properties
+    # Here we create a PDF which illustrates some properties of the dataset relating to the target values.
+    pdf_path = os.path.join(e.path, 'dataset.pdf')
+    with PdfPages(pdf_path) as pdf:
+        # First we are going to create a bar chart which will show for every target value, how many elements
+        # are actually annotated with a valid value (which is not None)
+        fig, ax = plt.subplots(ncols=1, nrows=1, figsize=(12, 10))
+        fig.suptitle(f'number of elements with valid target values\n'
+                     f'{dataset_length} elements total')
+        ax.set_xlabel('target property')
+        ax.set_ylabel('number of elements')
+        ax.set_xticks(range(len(TARGET_NAMES)))
+        ax.set_xticklabels(TARGET_NAMES)
+
+        target_values_map: t.Dict[str, t.List[float]] = {}
+        for i, name in enumerate(TARGET_NAMES):
+            target_values_map[name] = [v for g in dataset if (v := g['graph_labels'][i]) is not None]
+            num_valid = len(target_values_map[name])
+            ax.bar(
+                x=i,
+                height=num_valid,
+                color=COLOR_PRIMARY
+            )
+
+        pdf.savefig(fig)
+        plt.close(fig)
+
+        # We plot the value ranges and mean values for all the different target value distributions.
+        # This information will be important to decide on the model hyperparameters REGRESSION_REFERENCE
+        # and REGRESSION_LIMIT for all of the targets
+        fig, ax = plt.subplots(ncols=1, nrows=1, figsize=(12, 10))
+        fig.suptitle(f'target value ranges')
+        ax.set_xlabel('target value')
+        ax.set_ylabel('target property')
+        ax.set_yticks(range(len(TARGET_NAMES)))
+
+        y_labels = []
+        for i, name in enumerate(TARGET_NAMES):
+            values = target_values_map[name]
+            min_value = np.min(values)
+            max_value = np.max(values)
+            mean_value = np.mean(values)
+            ax.barh(
+                y=i,
+                left=min_value,
+                width=abs(max_value - min_value),
+                color=COLOR_PRIMARY,
+            )
+            ax.scatter(
+                y=i, x=mean_value,
+                color=COLOR_SECONDARY,
+                marker='d',
+            )
+            y_labels.append(f'{name}\n'
+                            f'({min_value:.1f} ; {mean_value:.1f} ; {max_value:.1f})')
+
+        ax.set_yticklabels(y_labels)
+        pdf.savefig(fig)
+        plt.close(fig)
 
     # This turns the list of graph dicts into the final form which we need for the training of the model:
     # keras RaggedTensors which contain all the graphs.
@@ -191,6 +277,7 @@ with Skippable(), (e := Experiment(base_path=BASE_PATH, namespace=NAMESPACE, glo
         regression_limits=REGRESSION_LIMITS,
         regression_reference=REGRESSION_REFERENCE,
         dropout_rate=DROPOUT_RATE,
+        final_dropout_rate=FINAL_DROPOUT,
         use_graph_attributes=HAS_GRAPH_ATTRIBUTES,
     )
 
@@ -231,9 +318,11 @@ with Skippable(), (e := Experiment(base_path=BASE_PATH, namespace=NAMESPACE, glo
     # -- EVALUATION OF TEST SET --
     e.info('evaluating test set...')
 
-    # first of all we need to pipe the entire test set through the model to get the target value predictions as well
-    # as the predicted node and edge importance explanations
+    # first of all we need to pipe the entire test set through the model to get the target value predictions
+    # as well as the predicted node and edge importance explanations
     out_pred, ni_pred, ei_pred = model(x_test)
+    out_pred = np.array(out_pred.numpy())
+    out_true = y_test[0]
 
     example_indices = random.sample(test_indices, k=int(EXAMPLES_RATIO * len(test_indices)))
     x_example, y_example, _, _ = process_graph_dataset(
@@ -248,9 +337,52 @@ with Skippable(), (e := Experiment(base_path=BASE_PATH, namespace=NAMESPACE, glo
     ei_example = ei_example.numpy()
 
     # -- VISUALIZATION OF RESULTS --
+    # The purpose of the following code is to create artifacts which document the results of the model
+    # training process in a human-readable way.
+
+    # ~ Visualizing the evaluation results
+    # This section generates a PDF which contains the R2 scores for all the different target values,
+    # calculated using the test set predictions.
+    # The PDF also visualizes the regression results using a regression "scatter" plot.
+    pdf_path = os.path.join(e.path, 'training.pdf')
+    with PdfPages(pdf_path) as pdf:
+        n_cols = len(TARGET_NAMES)
+        n_rows = 1
+
+        fig, rows = plt.subplots(ncols=n_cols, nrows=n_rows, figsize=(10 * n_cols, 10), squeeze=False)
+        for i, name in enumerate(TARGET_NAMES):
+            ax = rows[0][i]
+
+            values_true = []
+            values_pred = []
+            for value_true, value_pred in zip(out_true, out_pred):
+                if not np.isnan(value_true[i]) and not np.isnan(value_pred[i]):
+                    values_true.append(value_true[i])
+                    values_pred.append(value_pred[i])
+
+            plot_regression_fit(
+                values_true=values_true,
+                values_pred=values_pred,
+                ax=ax
+            )
+
+            r2 = r2_score(values_true, values_pred)
+            ax.set_title(f'"{name}"\n'
+                         f'r2: {r2:.2f}')
+
+        pdf.savefig(fig)
+        plt.close(fig)
+
+    # ~ Visualizing Examples
+    # This section will create a PDF which contains the illustrations of the predicted importance
+    # explanations of the model for a subset of the test set elements contained within the randomly
+    # sampled "example set".
     e.info(f'visualizing {len(example_indices)} example explanations...')
     pdf_path = os.path.join(e.path, 'examples.pdf')
     graph_list = [index_data_map[i]['metadata']['graph'] for i in example_indices]
+    # "create_importances_pdf" is a rather complex function, which, if given sufficient information about
+    # the dataset and the model predictions, will create a PDF file containing the visualized importance
+    # explanations using graph visualizations of the underlying VGD at a target location.
     create_importances_pdf(
         graph_list=graph_list,
         image_path_list=[index_data_map[i]['image_path'] for i in example_indices],